+++ validator/src/main/java/org/mustangproject/validator/XMLValidator.java
@@ -54,6 +54,7 @@ public class XMLValidator extends Validator {
 	int firedRules = 0;
 	int failedRules = 0;
 	boolean disableNotices = false;
+	boolean disableArithmeticCheck = false;
 	ISchematronResource aResSCH = null;
 
 
@@ -116,6 +117,13 @@ public class XMLValidator extends Validator {
 	 */
 	public void disableNotices() {
 		disableNotices = true;
+	}
+
+	/***
+	 * don't perform the arithmetic recalculation check in the validation report
+	 */
+	public void disableArithmeticCheck() {
+		disableArithmeticCheck = true;
 	}
 
 
@@ -406,7 +414,9 @@ public class XMLValidator extends Validator {
 						}
 					}
 				}
-				checkArithmetics(context);
+				if (!disableArithmeticCheck) {
+					checkArithmetics(context);
+				}
 
 
 			} catch (final IrrecoverableValidationError er) {